annotation text as label
2 changes made.  the first is to use the text of the annotation as the label instead of the generic "Annotation"

refresh annotation now refreshes all annotations, not just the new ones, so that labels are updated with modified text.

a side-effect of this is that if you change the label manually this change will be lost on refresh. a way around this would be to use a dictionary instead of lists for self.annotations and self.annotaitonsName  so you arent relying on list order to know which annotation goes with which name but this would be a bigger change in the code and i don't know if you want to go in that direction.